Simplify ParAccess to only have a single Record
This way it's up to the implementor of the Access whether
it should refer to immutable or immutable records.
Generally it gives us more flexibility, and allows us to have
a single implementation of a number of things, such as
create_par_iter.

However, any use of an Access must then assume mutable reference
semantics, i.e. two unsync accesses to a record cannot
happen at the same time, even if the actual Record allows this
(e.g. &T).